modify migration by changing quantity and min quantity attributes data type to unsigned int in order not to allow negative

+++ database/migrations/2020_01_28_111318_create_items_table.php
@@ -23,8 +23,8 @@ class CreateItemsTable extends Migration
             $table->uuid('item_unit_id')->nullable();
             $table->foreign('item_unit_id')->references('id')->on('item_units')->onUpdate('cascade')->onDelete('set null');
 
-            $table->float('min_quantity')->default(0);
-            $table->float('quantity')->default(0);
+            $table->unsignedInteger('min_quantity')->default(0);
+            $table->unsignedInteger('quantity')->default(0);
             
             $table->mediumText('desc')->nullable();
             $table->timestamps();
